requirements.txt
flask
flask-socketio
pycryptodome
gevent
gevent-websocket
